+++ api/src/main/java/io/strimzi/api/kafka/model/listener/LoadBalancerListenerBrokerOverride.java
@@ -28,6 +28,7 @@ public class LoadBalancerListenerBrokerOverride extends ExternalListenerBrokerOv
     private static final long serialVersionUID = 1L;
 
     private Map<String, String> dnsAnnotations = new HashMap<>(0);
+    private String loadBalancerIP;
 
     @Description("Annotations that will be added to the `Service` resources for individual brokers. " +
             "You can use this field to configure DNS providers such as External DNS.")
@@ -38,5 +39,17 @@ public class LoadBalancerListenerBrokerOverride extends ExternalListenerBrokerOv
 
     public void setDnsAnnotations(Map<String, String> dnsAnnotations) {
         this.dnsAnnotations = dnsAnnotations;
+    }
+
+    @Description("The loadbalancer is requested with the IP address specified in this field. " +
+            "This feature depends on whether the underlying cloud provider supports specifying the `loadBalancerIP` when a load balancer is created. " +
+            "This field is ignored if the cloud provider does not support the feature.")
+    @JsonInclude(JsonInclude.Include.NON_EMPTY)
+    public String getLoadBalancerIP() {
+        return loadBalancerIP;
+    }
+
+    public void setLoadBalancerIP(String loadBalancerIP) {
+        this.loadBalancerIP = loadBalancerIP;
     }
 }